Move tag command to edit command as its flag
