+++ cmd/root.go
@@ -33,7 +33,6 @@ func newRootCmd() *cobra.Command {
 
 	rootCmd.AddCommand(newEditCmd())
 	rootCmd.AddCommand(newNewCmd())
-	rootCmd.AddCommand(newTagCmd())
 	return rootCmd
 }
